Move scoresheet selector to its own component
